requirements.txt
numpy
scikit-learn
pyzmq
h5py
pyyaml
pyserial
